% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uniprot_info.R
\name{add_uniprot_info}
\alias{add_uniprot_info}
\title{add_uniprot_info}
\usage{
add_uniprot_info(listofproteins, taxon = NULL, database = NULL, tdrep = TRUE)
}
\arguments{
\item{tdrep}{}
}
\value{

}
\description{
add_uniprot_info
}
