% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_UP_database.R
\name{download_UP_database}
\alias{download_UP_database}
\title{download_UP_database}
\usage{
download_UP_database(taxon_number)
}
\arguments{
\item{taxon_number}{}
}
\value{

}
\description{
download_UP_database
}
