% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_data_path}
\alias{get_data_path}
\title{get_data_path}
\usage{
get_data_path(filedir, filename, extension)
}
\arguments{
\item{filedir}{Directory to search for file}

\item{filename}{Filename to get full path for}

\item{extension}{Vector of allowed file extensions}
}
\value{
A vector of full file paths
}
\description{
get_data_path
}
\examples{

}
