% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_locations_byfraction2}
\alias{get_locations_byfraction2}
\title{get_locations_byfraction2}
\usage{
get_locations_byfraction2(resultslist)
}
\arguments{
\item{resultslist}{}
}
\value{

}
\description{
get_locations_byfraction2
}
\examples{

}
