% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_locations_proteoform}
\alias{get_locations_proteoform}
\title{get_locations_proteoform}
\usage{
get_locations_proteoform(resultslist)
}
\arguments{
\item{resultslist}{}
}
\value{

}
\description{
get_locations_proteoform
}
