% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guppi.R
\name{guppi}
\alias{guppi}
\title{guppi}
\usage{
guppi(
  filedir,
  filename,
  taxon_number,
  outputdir,
  fdr = 0.01,
  saveOutput = TRUE,
  makeDashboard = FALSE,
  dashboardPath = NULL,
  usePB = FALSE
)
}
\value{
A variety of information sheets.
}
\description{
guppi
}
