% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{kickout}
\alias{kickout}
\title{kickout}
\usage{
kickout(list, allowed_ext = c("tdReport", "csv", "xlsx"))
}
\arguments{
\item{list}{A list of strings.}

\item{allowed_ext}{Vector of strings containing allowed file extensions.}
}
\value{

}
\description{
kickout
}
\examples{

}
