% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mods_allhits.R
\name{parse_mods_allhits}
\alias{parse_mods_allhits}
\title{parse_mods_allhits}
\usage{
parse_mods_allhits(tbl, modification = NULL)
}
\arguments{
\item{modification}{}
}
\value{

}
\description{
parse_mods_allhits
}
