% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tdreport_protein_full.R
\name{read_tdreport_protein_full}
\alias{read_tdreport_protein_full}
\title{read_tdreport_protein_full}
\usage{
read_tdreport_protein_full(tdreport, fdr_cutoff = 0.01)
}
\arguments{
\item{tdreport}{Full path to tdReport file}

\item{fdr_cutoff}{Cutoff for false detection rate, default 0.01}
}
\value{

}
\description{
read_tdreport_protein_full
}
