% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guppi.R
\name{guppi}
\alias{guppi}
\title{guppi}
\usage{
guppi(
  filedir,
  filename,
  taxonNumber,
  GOLocType = "bacteria",
  fractionAssignments = NULL,
  outputdir,
  fdr = 0.01,
  saveOutput = TRUE,
  makeDashboard = FALSE,
 
    dashboardPath = glue::glue("{outputdir}/report/{format(Sys.time(), '\%Y\%m\%d_\%H\%M\%S')}_dashboard.html")
)
}
\value{
A variety of information sheets.
}
\description{
guppi
}
