% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guppi.R
\name{guppi}
\alias{guppi}
\title{Get UniProt Protein Info (GUPPI)}
\usage{
guppi(
  filedir,
  filename,
  taxonNumber,
  GOLocType = "bacteria",
  fractionAssignments = NULL,
  outputdir,
  fdr = 0.01,
  saveOutput = TRUE,
  makeDashboard = FALSE,
 
    dashboardPath = glue::glue("{outputdir}/report/{format(Sys.time(), '\%Y\%m\%d_\%H\%M\%S')}_dashboard.html")
)
}
\arguments{
\item{filedir}{Character vector of the directory which contains the input file(s). Files can be at 
any level below the directory, subdirectories are checked recursively.}

\item{filename}{Character vector of the names of the tdReport files, including extension. 
Must match the name of the files exactly.}

\item{taxonNumber}{Numeric vector. NCBI taxon number to be used for getting UniProt data, 
e.g. 83333 for E. coli K12.}

\item{GOLocType}{Name of taxon to use for determination of subcellular locations. 
Acceptable values are "bacteria" and "eukaryota".}

\item{fractionAssignments}{Optional argument. A named list with names set to 
the fraction numbers of the input files and values set to the input file names. 
If left blank, GUPPI will attempt to assign fraction numbers automatically from 
file names in the tdReport.}

\item{outputdir}{Character vector. Directory to use for output, files will be 
saved to subdirectories. Directory will be created if it doesn't exist, as long 
as parent directory exists.}

\item{fdr}{False detection rate cutoff for protein and proteoform identifications. 
Defaults to 0.01 (1\%).}

\item{saveOutput}{Boolean value. Controls whether protein report, proteoform 
report, etc. are saved to output directory. Doesn't effect GUPPI report. 
Defaults to TRUE.}

\item{makeDashboard}{Boolean values. Controls whether GUPPI report is rendered 
to the output directory. Defaults to FALSE.}

\item{dashboardPath}{Full path and name of dashboard output (i.e. GUPPI report). 
Defaults to /report subdirectory of the output directory.}
}
\value{
This function is called for its side effects, i.e. the files in the output
directory.
}
\description{
guppi() analyzes a .tdReport file generated by TDPortal or ProSightPD and adds 
additional data from UniProt.
}
\examples{
\dontrun{
GUPPI::guppi(
   "C:/Users/David Butcher/TDReports",
   c(
      "20200420_Excellent_TDReport_01.tdReport",
      "20200420_Excellent_TDReport_02.tdReport"
   ),
   83333,
   GOLocType = "bacteria",
   fractionAssignments = NULL,
   outputdir = "C:/Users/David Butcher/guppi_output",
   fdr = 0.01,
   saveOutput = TRUE,
   makeDashboard = TRUE
)
}
}
